requirements.txt
pyrogram==2.0.106
pydub==0.25.1
python-dotenv==1.0.1
tgcrypto==1.2.5